refactor: 패키지 domain과 view로 분리
