+++ src/main/java/racingcar/domain/RacingResult.java
@@ -1,4 +1,4 @@
-package racingcar;
+package racingcar.domain;
 
 import java.util.ArrayList;
 import java.util.Collections;
